requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic>=2.9.0
presidio-analyzer==2.2.354
presidio-anonymizer==2.2.354
litellm>=1.34.21
pytest==8.1.1
httpx==0.27.0
python-dotenv==1.0.1
setuptools<70.0.0